+++ packages/headless/src/features/facets/category-facet-set/category-facet-utils.ts
@@ -4,7 +4,6 @@ import {
   logFacetDeselect,
   logFacetSelect,
 } from '../facet-set/facet-set-analytics-actions';
-import {FacetSelectionChangeMetadata} from '../facet-set/facet-set-analytics-actions-utils';
 
 type GenericCategoryFacetValue = CategoryFacetValueRequest | CategoryFacetValue;
 
@@ -42,7 +41,7 @@ export const getAnalyticsActionForCategoryFacetToggleSelect = (
   facetId: string,
   selection: CategoryFacetValue
 ) => {
-  const payload: FacetSelectionChangeMetadata = {
+  const payload = {
     facetId,
     facetValue: selection.value,
   };