+++ packages/headless/src/features/facets/facet-api/request.ts
@@ -68,11 +68,15 @@ export interface Expandable {
   isFieldExpanded: boolean;
 }
 
-export interface Type<
-  T extends 'specific' | 'dateRange' | 'numericalRange' | 'hierarchical'
-> {
+export interface Type<T extends FacetType> {
   type: T;
 }
+
+export type FacetType =
+  | 'specific'
+  | 'dateRange'
+  | 'numericalRange'
+  | 'hierarchical';
 
 export interface SortCriteria<
   T extends